Factor sub transitions out of state_transition
This aims to keep the global view of the state machine clear;
instead of one heavily branching `state_transition` function, it introduces
several sub transition methods, that are only applicable to a certain (machine)
state of the `State` object. Asessment of the (for lack of a better word)
"stable" states is provided by `@property`s of the `State` object.

This will allow for finer unit tests, where we only need to test the sub
transitions and can more easily reason about the sub state of the machine.