Fix the unhappy-case stress test
The stress test deadlocked on crashing nodes, because `stop()`ing could
lead the matrix message handler to wait on itself. This change makes
sure that crash and restart are executed in their own greenlet, so this
is independent from the message handler.

Getting rid of the dedicated `_monitor` also allows to avoid crashing
while restarting, which made the test nodes unresponsive and
unrecoverable.