Transaction handled by inidividual greenlets
To avoid starvation of the AlarmTask, the handling of raiden events must
be done in separated threads.

Ideally transactions would be added to a queue, allowing them to be
batched, however this requires support from the JSON RPC client, which
is not available, so currenlty we are implicitely queing the
transactions in greenlets and the order in which the transactions are
mined.

For the off-chain messages the queues are in place, and it's sufficient
to just queue them using send_async. Note, however, it's very important
for the messages to be queue in order, for this reason greenlets are
/not/ used when there is an event for a protocol message. If greenlets
where to be used, the order of messages could change, and if a message
with a higher nonce got added to the queue first, then the queue would
stuck.